# [Xamarin.Android.Build.Tasks] use System.Reflection.Metadata for ResolveLibraryProjectImports