--- a/src/Xamarin.Android.Build.Tasks/Tasks/ResolveLibraryProjectImports.cs
+++ b/src/Xamarin.Android.Build.Tasks/Tasks/ResolveLibraryProjectImports.cs
@@ -3,14 +3,12 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Xml.Linq;
-using Mono.Cecil;
 using Microsoft.Build.Utilities;
 using Microsoft.Build.Framework;
 using Xamarin.Tools.Zip;
-
-using Java.Interop.Tools.Cecil;
-
 using Xamarin.Android.Tools;
+using System.Reflection.PortableExecutable;
+using System.Reflection.Metadata;
 
 namespace Xamarin.Android.Tasks
 {
@@ -84,13 +82,11 @@ namespace Xamarin.Android.Tasks
 			var resolvedEnvironmentFiles      = new List<ITaskItem> ();
 
 			assemblyMap.Load (AssemblyIdentityMapFile);
-			using (var resolver = new DirectoryAssemblyResolver (this.CreateTaskLogger (), loadDebugSymbols: false)) {
-				try {
-					Extract (resolver, jars, resolvedResourceDirectories, resolvedAssetDirectories, resolvedEnvironmentFiles);
-				} catch (ZipIOException ex) {
-					Log.LogCodedError ("XA1004", ex.Message);
-					Log.LogDebugMessage (ex.ToString ());
-				}
+			try {
+				Extract (jars, resolvedResourceDirectories, resolvedAssetDirectories, resolvedEnvironmentFiles);
+			} catch (ZipIOException ex) {
+				Log.LogCodedError ("XA1004", ex.Message);
+				Log.LogDebugMessage (ex.ToString ());
 			}
 
 			Jars                        = jars.Values.ToArray ();
@@ -143,7 +139,6 @@ namespace Xamarin.Android.Tasks
 		// Extracts library project contents under e.g. obj/Debug/[__library_projects__/*.jar | res/*/*]
 		// Extracts library project contents under e.g. obj/Debug/[lp/*.jar | res/*/*]
 		void Extract (
-				DirectoryAssemblyResolver res,
 				IDictionary<string, ITaskItem> jars,
 				ICollection<ITaskItem> resolvedResourceDirectories,
 				ICollection<ITaskItem> resolvedAssetDirectories,
@@ -157,9 +152,6 @@ namespace Xamarin.Android.Tasks
 			}
 			var outdir = Path.GetFullPath (OutputImportDirectory);
 			Directory.CreateDirectory (outdir);
-
-			foreach (var assembly in Assemblies)
-				res.Load (assembly.ItemSpec);
 
 			bool skip;
 			foreach (var assemblyItem in Assemblies) {
@@ -224,100 +216,98 @@ namespace Xamarin.Android.Tasks
 
 				Log.LogDebugMessage ($"Refreshing {assemblyFileName}.dll");
 
-				var assembly = res.GetAssembly (assemblyPath);
-				foreach (var mod in assembly.Modules) {
-					// android environment files
-					foreach (var envtxt in mod.Resources
-							.Where (r => r.Name.StartsWith ("__AndroidEnvironment__", StringComparison.OrdinalIgnoreCase))
-							.Where (r => r is EmbeddedResource)
-							.Cast<EmbeddedResource> ()) {
-						var outFile = Path.Combine (outDirForDll, envtxt.Name);
-						using (var stream = envtxt.GetResourceStream ()) {
-							updated |= MonoAndroidHelper.CopyIfStreamChanged (stream, outFile);
-						}
-						resolvedEnvironments.Add (new TaskItem (Path.GetFullPath (outFile), new Dictionary<string, string> {
-							{ OriginalFile, assemblyPath }
-						}));
-					}
+				using (var pe = new PEReader (File.OpenRead (assemblyPath))) {
 
-					// embedded jars (EmbeddedJar, EmbeddedReferenceJar)
-					var resjars = mod.Resources
-						.Where (r => r.Name.EndsWith (".jar", StringComparison.InvariantCultureIgnoreCase))
-						.Select (r => (EmbeddedResource) r);
-					foreach (var resjar in resjars) {
-						using (var stream = resjar.GetResourceStream ()) {
-							AddJar (jars, importsDir, resjar.Name, assemblyPath);
-							updated |= MonoAndroidHelper.CopyIfStreamChanged (stream, Path.Combine (importsDir, resjar.Name));
-						}
-					}
+					var reader = pe.GetMetadataReader ();
 
-					var libzip = mod.Resources.FirstOrDefault (r => r.Name == "__AndroidNativeLibraries__.zip") as EmbeddedResource;
-					if (libzip != null) {
-						List<string> files = new List<string> ();
-						using (var stream = libzip.GetResourceStream ())
-						using (var zip = Xamarin.Tools.Zip.ZipArchive.Open (stream)) {
-							try {
-								updated |= Files.ExtractAll (zip, nativeimportsDir, modifyCallback: (entryFullName) => {
-									files.Add (Path.GetFullPath (Path.Combine (nativeimportsDir, entryFullName)));
-									return entryFullName
-										.Replace ("native_library_imports\\", "")
-										.Replace ("native_library_imports/", "");
-								}, deleteCallback: (fileToDelete) => {
-									return !files.Contains (fileToDelete);
-								});
-							} catch (PathTooLongException ex) {
-								Log.LogCodedError ("XA4303", $"Error extracting resources from \"{assemblyPath}\": {ex}");
-								return;
-							} catch (NotSupportedException ex) {
-								Log.LogCodedError ("XA4303", $"Error extracting resources from \"{assemblyPath}\": {ex}");
-								return;
+					foreach (var handle in reader.ManifestResources) {
+						var resource = reader.GetManifestResource (handle);
+						string name = reader.GetString (resource.Name);
+
+						// android environment files
+						if (name.StartsWith ("__AndroidEnvironment__", StringComparison.OrdinalIgnoreCase)) {
+							var outFile = Path.Combine (outDirForDll, name);
+							using (var stream = pe.GetEmbeddedResourceStream (resource)) {
+								updated |= MonoAndroidHelper.CopyIfStreamChanged (stream, outFile);
 							}
-						}
-					}
-
-					// embedded AndroidResourceLibrary archive
-					var reszip = mod.Resources.FirstOrDefault (r => r.Name == "__AndroidLibraryProjects__.zip") as EmbeddedResource;
-					if (reszip != null) {
-						// temporarily extracted directory will look like:
-						//    __library_projects__/[dllname]/[library_project_imports | jlibs]/bin
-						using (var stream = reszip.GetResourceStream ())
-						using (var zip = Xamarin.Tools.Zip.ZipArchive.Open (stream)) {
-							try {
-								updated |= Files.ExtractAll (zip, importsDir, modifyCallback: (entryFullName) => {
-									var path = entryFullName
-										.Replace ("library_project_imports\\","")
-										.Replace ("library_project_imports/", "");
-									if (path.EndsWith (".jar", StringComparison.OrdinalIgnoreCase)) {
-										AddJar (jars, importsDir, path, assemblyPath);
-									}
-									return path;
-								}, deleteCallback: (fileToDelete) => {
-									return !jars.ContainsKey (fileToDelete);
-								});
-							} catch (PathTooLongException ex) {
-								Log.LogCodedError ("XA4303", $"Error extracting resources from \"{assemblyPath}\": {ex}");
-								return;
-							} catch (NotSupportedException ex) {
-								Log.LogCodedError ("XA4303", $"Error extracting resources from \"{assemblyPath}\": {ex}");
-								return;
-							}
-						}
-
-						// We used to *copy* the resources to overwrite other resources,
-						// which resulted in missing resource issue.
-						// Here we replaced copy with use of '-S' option and made it to work.
-						if (Directory.Exists (resDir)) {
-							var taskItem = new TaskItem (Path.GetFullPath (resDir), new Dictionary<string, string> {
-								{ OriginalFile, assemblyPath }
-							});
-							if (bool.TryParse (assemblyItem.GetMetadata (AndroidSkipResourceProcessing), out skip) && skip)
-								taskItem.SetMetadata (AndroidSkipResourceProcessing, "True");
-							resolvedResourceDirectories.Add (taskItem);
-						}
-						if (Directory.Exists (assetsDir))
-							resolvedAssetDirectories.Add (new TaskItem (Path.GetFullPath (assetsDir), new Dictionary<string, string> {
+							resolvedEnvironments.Add (new TaskItem (Path.GetFullPath (outFile), new Dictionary<string, string> {
 								{ OriginalFile, assemblyPath }
 							}));
+						}
+						// embedded jars (EmbeddedJar, EmbeddedReferenceJar)
+						else if (name.EndsWith (".jar", StringComparison.InvariantCultureIgnoreCase)) {
+							using (var stream = pe.GetEmbeddedResourceStream (resource)) {
+								AddJar (jars, importsDir, name, assemblyPath);
+								updated |= MonoAndroidHelper.CopyIfStreamChanged (stream, Path.Combine (importsDir, name));
+							}
+						}
+						// embedded native libraries
+						else if (name == "__AndroidNativeLibraries__.zip") {
+							List<string> files = new List<string> ();
+							using (var stream = pe.GetEmbeddedResourceStream (resource))
+							using (var zip = Xamarin.Tools.Zip.ZipArchive.Open (stream)) {
+								try {
+									updated |= Files.ExtractAll (zip, nativeimportsDir, modifyCallback: (entryFullName) => {
+										files.Add (Path.GetFullPath (Path.Combine (nativeimportsDir, entryFullName)));
+										return entryFullName
+											.Replace ("native_library_imports\\", "")
+											.Replace ("native_library_imports/", "");
+									}, deleteCallback: (fileToDelete) => {
+										return !files.Contains (fileToDelete);
+									});
+								} catch (PathTooLongException ex) {
+									Log.LogCodedError ("XA4303", $"Error extracting resources from \"{assemblyPath}\": {ex}");
+									return;
+								} catch (NotSupportedException ex) {
+									Log.LogCodedError ("XA4303", $"Error extracting resources from \"{assemblyPath}\": {ex}");
+									return;
+								}
+							}
+						}
+						// embedded AndroidResourceLibrary archive
+						else if (name == "__AndroidLibraryProjects__.zip") {
+							// temporarily extracted directory will look like:
+							//    __library_projects__/[dllname]/[library_project_imports | jlibs]/bin
+							using (var stream = pe.GetEmbeddedResourceStream (resource))
+							using (var zip = Xamarin.Tools.Zip.ZipArchive.Open (stream)) {
+								try {
+									updated |= Files.ExtractAll (zip, importsDir, modifyCallback: (entryFullName) => {
+										var path = entryFullName
+											.Replace ("library_project_imports\\", "")
+											.Replace ("library_project_imports/", "");
+										if (path.EndsWith (".jar", StringComparison.OrdinalIgnoreCase)) {
+											AddJar (jars, importsDir, path, assemblyPath);
+										}
+										return path;
+									}, deleteCallback: (fileToDelete) => {
+										return !jars.ContainsKey (fileToDelete);
+									});
+								} catch (PathTooLongException ex) {
+									Log.LogCodedError ("XA4303", $"Error extracting resources from \"{assemblyPath}\": {ex}");
+									return;
+								} catch (NotSupportedException ex) {
+									Log.LogCodedError ("XA4303", $"Error extracting resources from \"{assemblyPath}\": {ex}");
+									return;
+								}
+							}
+
+							// We used to *copy* the resources to overwrite other resources,
+							// which resulted in missing resource issue.
+							// Here we replaced copy with use of '-S' option and made it to work.
+							if (Directory.Exists (resDir)) {
+								var taskItem = new TaskItem (Path.GetFullPath (resDir), new Dictionary<string, string> {
+									{ OriginalFile, assemblyPath }
+								});
+								if (bool.TryParse (assemblyItem.GetMetadata (AndroidSkipResourceProcessing), out skip) && skip)
+									taskItem.SetMetadata (AndroidSkipResourceProcessing, "True");
+								resolvedResourceDirectories.Add (taskItem);
+							}
+							if (Directory.Exists (assetsDir)) {
+								resolvedAssetDirectories.Add (new TaskItem (Path.GetFullPath (assetsDir), new Dictionary<string, string> {
+									{ OriginalFile, assemblyPath }
+								}));
+							}
+						}
 					}
 				}