I wrote the code for this a long time ago, but never opened a PR
because when timing the difference, I was seeing *no* improvement.

I even wrote a benchmark, that did clearly show an improvement:

https://github.com/jonathanpeppers/Benchmarks/blob/38ffb60421604b15eef38170ad9b12a2049ac3f1/Benchmarks/Cecil.cs

         Method |      Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
    ----------- |----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
            SRM |  8.669 ms | 0.0385 ms | 0.0322 ms |   1000.0000 |    671.8750 |    656.2500 |             5.51 MB |
     Mono.Cecil | 24.006 ms | 0.1879 ms | 0.1758 ms |   1906.2500 |   1718.7500 |    781.2500 |            39.95 MB |

Then last week I saw this line in the diff:

    var assemblyLastWrite = new FileInfo (assemblyPath).LastWriteTimeUtc;

Doh! This was leftover from something else, I'm not sure!

After removing that, this *does* in fact help performance:

    Before:
    4814 ms  ResolveLibraryProjectImports               1 calls
    After:
    4502 ms  ResolveLibraryProjectImports               1 calls

This was the Xamarin.Forms project in this repo. I think this will
help by around 300ms any time the `<ResolveLibraryProjectImports/>`
MSBuild task needs to run.

This seems like an easy-win. The only drawback is the addition of:

    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>

The `unsafe` keyword is needed for accessing `EmbeddedResource` with
System.Reflection.Metadata.